optimized user-function calling, for loops, @variables, ifs with jumps
user defined functions now default to @inline macro-like compilation, which generally minimizes execution cycles, though potentially yields much longer programs if functions are long and repeatedly called.  If a variable is passed as an arg, and an inline function modifies that arg, that will change the variable.  This macro-like behavior can be useful, or can be blocked by passing variable+0 to the function rather than just the naked variable.

More traditional out-of-line function compilation can be selected by preceding the def with the @function decorator.  This now passes arguments and return_lines via variables rather than the stack.  Argument names are mangled to avoid collisions with lookalike names outside the function.  Out-of-line compilation may be preferable if program length becomes an issue, or to simplify some jumps in/out of the function, e.g., if you want to call it recursively (though there is no stack to track earlier callers, so you'll need explicit jumps to escape the recursion).

Since the stack is no longer used it has been removed.

Function body instructions are now stored with the function, and appended either inline at call locations, or at end of program.  The main program is now stored in compiler._main, and compiler._current indicates whether instructions are currently being added to ._main or to some function's definition.

Added as_target (analog of as_value for the target of a set instruction) and and as_name (used by both as_value and as_target), needed for processing argument occurences within functions.  Incorporated into those processing of special @variables like Unit -> @unit, This -> @this

For loops now use the new label system for line-numbering, needed so they can work in functions.

Added some optimization checks on function and if-then flow of control to help eliminate the pointless inclusion of instructions that would never be executed because the preceding instruction jumped away.

Many tests have been updated to accomodate elimination of stack (which changed line numbering by 1 on everything) and ways in which functions, for-loops, and if-thens have now been optimized.